feat: 🎸 addressed pr comments
