+++ src/components/core/buttons/index.ts
@@ -1,4 +1,3 @@
 export * from './action-button';
 export * from './link-button';
-export * from './plug-button';
 export * from './icon-action-button';
